Ds_Virtual_space_micro/services/auth-service/requirements.txt
flask==3.0.3
flask-cors==5.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
supabase==2.0.0
redis==5.0.8
gunicorn==23.0.0